AI_Article_generate/requirements.txt
streamlit==1.51.0
google-genai==1.49.0
dotenv==0.9.9